Split Lettuce attribute getter tests
