+++ instrumentation/lettuce/lettuce-4.0/javaagent-unit-tests/src/test/java/io/opentelemetry/javaagent/instrumentation/lettuce/v4_0/LettuceBatchAttributesGetterTest.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.lettuce.v4_0;
+
+import static io.opentelemetry.instrumentation.api.internal.SemconvStability.emitStableDatabaseSemconv;
+import static java.util.Collections.singletonList;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.lambdaworks.redis.codec.Utf8StringCodec;
+import com.lambdaworks.redis.output.StatusOutput;
+import com.lambdaworks.redis.protocol.Command;
+import com.lambdaworks.redis.protocol.CommandType;
+import com.lambdaworks.redis.protocol.RedisCommand;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.RedisServerTarget;
+import java.net.InetSocketAddress;
+import org.junit.jupiter.api.Test;
+
+class LettuceBatchAttributesGetterTest {
+
+  private static final InetSocketAddress SELECTED_ADDRESS =
+      InetSocketAddress.createUnresolved("selected-node", 6380);
+
+  @Test
+  void batchUsesConfiguredTargetOnlyForStableSemconv() {
+    LettuceBatchRequest request =
+        LettuceBatchRequest.create(
+            singletonList(command()),
+            SELECTED_ADDRESS,
+            null,
+            RedisServerTarget.ofEndpoint("configured-node:6379"));
+
+    LettuceBatchAttributesGetter getter = new LettuceBatchAttributesGetter();
+
+    assertThat(getter.getServerAddress(request))
+        .isEqualTo(emitStableDatabaseSemconv() ? "configured-node" : "selected-node");
+    assertThat(getter.getServerPort(request)).isEqualTo(emitStableDatabaseSemconv() ? null : 6380);
+  }
+
+  private static RedisCommand<String, String, String> command() {
+    Utf8StringCodec codec = new Utf8StringCodec();
+    return new Command<>(CommandType.GET, new StatusOutput<>(codec));
+  }
+}
