+++ instrumentation/lettuce/lettuce-4.0/javaagent-unit-tests/src/test/java/io/opentelemetry/javaagent/instrumentation/lettuce/v4_0/LettuceDbAttributesGetterTest.java
@@ -6,7 +6,6 @@
 package io.opentelemetry.javaagent.instrumentation.lettuce.v4_0;
 
 import static io.opentelemetry.instrumentation.api.internal.SemconvStability.emitStableDatabaseSemconv;
-import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lambdaworks.redis.codec.Utf8StringCodec;
@@ -18,7 +17,7 @@ import io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.RedisS
 import java.net.InetSocketAddress;
 import org.junit.jupiter.api.Test;
 
-class LettuceAttributesGetterTest {
+class LettuceDbAttributesGetterTest {
 
   private static final InetSocketAddress SELECTED_ADDRESS =
       InetSocketAddress.createUnresolved("selected-node", 6380);
@@ -47,22 +46,6 @@ class LettuceAttributesGetterTest {
     assertThat(getter.getServerAddress(command))
         .isEqualTo(emitStableDatabaseSemconv() ? null : "selected-node");
     assertThat(getter.getServerPort(command)).isEqualTo(emitStableDatabaseSemconv() ? null : 6380);
-  }
-
-  @Test
-  void batchUsesConfiguredTargetOnlyForStableSemconv() {
-    LettuceBatchRequest request =
-        LettuceBatchRequest.create(
-            singletonList(command()),
-            SELECTED_ADDRESS,
-            null,
-            RedisServerTarget.ofEndpoint("configured-node:6379"));
-
-    LettuceBatchAttributesGetter getter = new LettuceBatchAttributesGetter();
-
-    assertThat(getter.getServerAddress(request))
-        .isEqualTo(emitStableDatabaseSemconv() ? "configured-node" : "selected-node");
-    assertThat(getter.getServerPort(request)).isEqualTo(emitStableDatabaseSemconv() ? null : 6380);
   }
 
   private static RedisCommand<String, String, String> command() {
